app/requirements.txt
python-dotenv
supabase
pandas
requests
beautifulsoup4
geopy
unidecode
Flask
scikit-learn
pyjwt
gunicorn
google-genai